added validation properties and custom IValidatableObject also implemented ValidateByValidateObject and ValidateByTryValidateObject

+++ CryptoMonitor/CryptoMonitor.BL/IManager.cs
@@ -31,7 +31,7 @@ public interface IManager
     void AddUserReview(
         string userName, 
         string comment, 
-        double rating, 
+        int rating, 
         Exchange exchange
         );
 }
